fix only admin can create/delete books, remove @Entity from model User, delete validation in service, create GetBooksByIdRequest, delete BookServiceException

+++ src/main/java/id/ac/ui/cs/advprog/admin/services/BookService.java
@@ -1,27 +1,32 @@
 package id.ac.ui.cs.advprog.admin.services;
 
-import java.util.Arrays;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
 
+import id.ac.ui.cs.advprog.admin.dto.GetBooksByIdRequestDto;
 import id.ac.ui.cs.advprog.admin.dto.PatchBookRequestDto;
 import id.ac.ui.cs.advprog.admin.dto.PostBookRequestDto;
-import id.ac.ui.cs.advprog.admin.enums.BookServiceException;
-import org.apache.tomcat.util.http.parser.Authorization;
+
+import id.ac.ui.cs.advprog.admin.dto.PostBookResponseDto;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.http.HttpStatus;
+import org.springframework.web.client.HttpClientErrorException;
 import org.springframework.web.client.HttpStatusCodeException;
 import org.springframework.http.MediaType;
 import org.springframework.stereotype.Service;
 import org.springframework.web.client.RestClient;
-import org.springframework.web.client.RestClientException;
 import org.springframework.web.util.UriComponentsBuilder;
 
 import id.ac.ui.cs.advprog.admin.models.Book;
 
 @Service
 public class BookService {
+
+    private static final Logger logger = LoggerFactory.getLogger(BookService.class);
 
     @Value("${api.book-host}")
     public String bookHost;
@@ -35,60 +40,70 @@ public class BookService {
         this.restClient = restClient;
     }
 
-    public Book save(Book book, String authHeader) {
+    public PostBookResponseDto save(PostBookRequestDto bookDto)
+            throws HttpStatusCodeException
+    {
         try {
             return restClient.post()
                     .uri(bookHost + book_url)
-                    .header("Authorization", authHeader)
                     .contentType(MediaType.APPLICATION_JSON)
-                    .body(book)
+                    .body(bookDto)
                     .retrieve()
-                    .body(Book.class);
-        } catch (HttpStatusCodeException e) {
-            throw new BookServiceException("Error saving book: " + e.getMessage(), e);
+                    .body(PostBookResponseDto.class);
+        } catch (HttpStatusCodeException exception) {
+            logger.error(exception.getMessage(), exception);
+            throw exception;
         }
     }
 
-    //This method still needs work to do since it deletes then save the updated book, so the id is different (not intented)
-    public void update(UUID bookId, PatchBookRequestDto dto) {
+    public void updateById(UUID bookId, PatchBookRequestDto dto)
+            throws HttpStatusCodeException
+    {
         try {
             restClient.patch()
                     .uri(bookHost + book_url2 + bookId)
                     .contentType(MediaType.APPLICATION_JSON)
                     .body(dto)
-                    .retrieve()
-                    .toBodilessEntity();
-        } catch (HttpStatusCodeException e) {
-            throw new BookServiceException("Error saving book: " + e.getMessage(), e);
+                    .retrieve();
+        } catch (HttpStatusCodeException exception) {
+            logger.error(exception.getMessage(), exception);
+            throw exception;
         }
     }
 
-    public Optional<Book> findById(UUID bookId) {
+    public Book findById(UUID bookId) throws HttpStatusCodeException {
         try {
-            Book book = restClient.get()
+            return restClient.get()
                     .uri(bookHost + book_url2 + bookId)
                     .retrieve()
                     .body(Book.class);
-            return Optional.ofNullable(book);
-        } catch (HttpStatusCodeException e) {
-            throw new BookServiceException("Error finding book by ID: " + e.getMessage(), e);
+        } catch (HttpStatusCodeException exception) {
+            logger.error(exception.getMessage(), exception);
+            throw exception;
         }
     }
 
-    public void deleteById(UUID bookId) {
+    public void deleteById(UUID bookId) throws HttpStatusCodeException {
         try {
-            restClient.delete()
-                    .uri(bookHost + book_url2 + bookId)
-                    .retrieve();
-        } catch (HttpStatusCodeException e) {
-            throw new BookServiceException("Error deleting book: " + e.getMessage(), e);
+            Book book = findById(bookId);
+            if (book.getDownloadCount() == 0) {
+                restClient.delete()
+                        .uri(bookHost + book_url2 + bookId)
+                        .retrieve();
+            } else {
+                logger.warn("Attempted to delete a book with downloads > 0: {}", bookId);
+                throw new HttpClientErrorException(HttpStatus.FORBIDDEN, "Cannot delete a book with downloads > 0");
+            }
+        } catch (HttpStatusCodeException exception) {
+            logger.error(exception.getMessage(), exception);
+            throw exception;
         }
     }
 
     public List<Book> findAll(
-        Optional<String> author, Optional<String> title,
-        Optional<String> sortBy, Optional<String> orderBy)
-    throws RestClientException
+            Optional<String> author, Optional<String> title,
+            Optional<String> sortBy, Optional<String> orderBy)
+            throws HttpStatusCodeException
     {
         UriComponentsBuilder uriBuilder = UriComponentsBuilder.fromHttpUrl(bookHost + book_url);
 
@@ -105,11 +120,34 @@ public class BookService {
             uriBuilder.queryParam("orderBy", orderBy);
         }
 
-        Book[] books = restClient.get()
-                .uri(uriBuilder.toUriString())
-                .retrieve()
-                .body(Book[].class);
+        Book[] books;
+        try {
+            books = restClient.get()
+                    .uri(uriBuilder.toUriString())
+                    .retrieve()
+                    .body(Book[].class);
+        } catch (HttpStatusCodeException exception) {
+            logger.error(exception.getMessage(), exception);
+            throw exception;
+        }
 
-        return Arrays.asList(books);
+        return List.of(books);
+    }
+
+    public List<Book> findAllById(GetBooksByIdRequestDto dto) throws HttpStatusCodeException {
+        Book[] books;
+        try {
+            books = restClient.post()
+                    .uri(bookHost + book_url + "/get-multiple")
+                    .accept(MediaType.APPLICATION_JSON)
+                    .body(dto)
+                    .retrieve()
+                    .body(Book[].class);
+        } catch (HttpStatusCodeException exception) {
+            logger.error(exception.getMessage(), exception);
+            throw exception;
+        }
+
+        return List.of(books);
     }
 }
